+++ src/bgw/job.h
@@ -45,7 +45,7 @@ extern TSDLLEXPORT bool ts_bgw_job_update_by_id(int32 job_id, BgwJob *job);
 extern TSDLLEXPORT int32 ts_bgw_job_insert_relation(
 	Name application_name, Interval *schedule_interval, Interval *max_runtime, int32 max_retries,
 	Interval *retry_period, Name proc_schema, Name proc_name, Name check_schema, Name check_name,
-	Name owner, bool scheduled, bool fixed_schedule, int32 hypertable_id, Jsonb *config,
+	Oid owner, bool scheduled, bool fixed_schedule, int32 hypertable_id, Jsonb *config,
 	TimestampTz initial_start, const char *timezone);
 extern TSDLLEXPORT void ts_bgw_job_permission_check(BgwJob *job);